+++ applications/vessel-history/src/data/config.ts
@@ -7,11 +7,23 @@ export const WORKSPACE_ENV =
   (process.env.NODE_ENV as WorkspaceEnv) ||
   'production'
 
+export const FLY_EFFECTS = {
+  noFly: 0, // just change the center
+  softFly: 1, // fly to without effects
+  fly: 2, // nice fly to moving the camera
+}
+// enable / disable the effect for switch events in the map
+export const ENABLE_FLYTO = FLY_EFFECTS.softFly //maybe we can move this to the settings
 export const DEBUG_MODE =
   ((process.env.REACT_APP_DEBUG_MODE ?? false) as boolean) ||
   (WORKSPACE_ENV === 'production' ? false : true)
 
 export const FULL_SUFIX = 'full'
+export const WORKSPACE_START_DATE = new Date()
+WORKSPACE_START_DATE.setMonth(WORKSPACE_START_DATE.getMonth() - 12)
+
+export const FIRST_YEAR_OF_DATA = 2012
+export const CURRENT_YEAR = new Date().getFullYear()
 
 export const DEFAULT_WORKSPACE: AppState = {
   zoom: 3,
@@ -26,7 +38,7 @@ export const DEFAULT_WORKSPACE: AppState = {
   fishingPositions: 15,
   longitude: -35.97144,
   project: '1',
-  start: '2017-12-01T00:00:00.000Z',
+  start: WORKSPACE_START_DATE.toISOString(),
   end: DateTime.utc().toISO(),
   timebarMode: 'speed',
   filterMode: 'speed',
@@ -34,6 +46,8 @@ export const DEFAULT_WORKSPACE: AppState = {
   maxDistanceFromPort: 10000,
   importView: false,
   satellite: '',
+  availableStart: new Date(Date.UTC(FIRST_YEAR_OF_DATA, 0, 1)).toISOString(),
+  availableEnd: new Date(Date.UTC(CURRENT_YEAR, 11, 31)).toISOString(),
 }
 
 export const DEFAULT_VIEWPORT = {