+++ applications/vessel-history/src/features/home/LanguageToggle.tsx
@@ -1,0 +1,46 @@
+import { Fragment, useState, useCallback } from 'react'
+import { useTranslation } from 'react-i18next'
+import cx from 'classnames'
+import { IconButton } from '@globalfishingwatch/ui-components'
+import { Locale } from 'types'
+import { LocaleLabels } from 'features/i18n/i18n'
+import styles from './LanguageToggle.module.css'
+
+const LanguageToggle: React.FC = () => {
+  const { i18n } = useTranslation()
+  const [open, setOpen] = useState(false)
+  const toggleLanguage = useCallback(
+    (lang: Locale) => {
+      i18n.changeLanguage(lang)
+      setOpen(false)
+    },
+    [setOpen, i18n]
+  )
+  const onOpen = useCallback(() => {
+    setOpen(!open)
+  }, [open, setOpen])
+  return (
+    <Fragment>
+      <IconButton type="default" size="default" className={styles.button} onClick={onOpen}>
+        {i18n.language}
+        {open && (
+          <ul className={styles.list}>
+            {LocaleLabels.map(({ id, label }) => (
+              <li
+                key={id}
+                className={cx(styles.language, {
+                  [styles.currentLanguage]: i18n.language === id,
+                })}
+                onClick={() => toggleLanguage(id)}
+              >
+                {label}
+              </li>
+            ))}
+          </ul>
+        )}
+      </IconButton>
+    </Fragment>
+  )
+}
+
+export default LanguageToggle