+++ applications/temporalgrid-demo/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @globalfishingwatchapp/temporalgrid-demo
 
+## 0.7.131
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@9.2.2
+
+## 0.7.130
+
+### Patch Changes
+
+- Updated dependencies [28d7d17c]
+- Updated dependencies [567ee29e]
+  - @globalfishingwatch/layer-composer@9.2.1
+  - @globalfishingwatch/timebar@2.0.8
+  - @globalfishingwatch/react-hooks@14.0.0
+
+## 0.7.129
+
+### Patch Changes
+
+- Updated dependencies [72ca7097]
+  - @globalfishingwatch/timebar@2.0.7
+
+## 0.7.128
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@9.2.0
+
 ## 0.7.127
 
 ### Patch Changes