+++ applications/vessel-history/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @globalfishingwatchapp/vessel-history
 
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [b706fcb3]
+  - @globalfishingwatch/api-client@3.1.1
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@9.2.2
+
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [998b9274]
+  - @globalfishingwatch/api-types@5.3.1
+
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [28d7d17c]
+- Updated dependencies [fa078d98]
+- Updated dependencies [25440a9f]
+- Updated dependencies [567ee29e]
+  - @globalfishingwatch/layer-composer@9.2.1
+  - @globalfishingwatch/dataviews-client@13.0.0
+  - @globalfishingwatch/ui-components@11.0.0
+  - @globalfishingwatch/timebar@2.0.8
+  - @globalfishingwatch/react-hooks@14.0.0
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [5b01cd38]
+  - @globalfishingwatch/api-types@5.3.0
+
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@9.2.0
+
 ## 1.0.3
 
 ### Patch Changes