+++ packages/api-client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/api-client
 
+## 3.1.1
+
+### Patch Changes
+
+- b706fcb3: do not send empty multi select array
+
+## 3.1.0
+
+### Minor Changes
+
+- 8396e437: Move advanced query logic to api-client
+
 ## 3.0.0
 
 ### Patch Changes